django_zodb/cursor.py
from django.db import DatabaseError
from BTrees.OOBTree import OOBTree
from zope.catalog.catalog import Catalog
from zope.catalog.field import FieldIndex
import transaction

class ZODBCursor:
    def __init__(self, connection):
        self.connection = connection
        self.lastrowid = None  # Initialize lastrowid attribute
        self.rowcount = -1  # Initialize rowcount attribute
        if 'catalog' not in self.connection.root:
            self.connection.root['catalog'] = Catalog()
            transaction.commit()

    def execute(self, sql, params=None):
        # Convert sql to string if it's not already
        sql = str(sql)
        # Implement interaction with ZODB here
        print(f"Executing SQL: {sql} with params: {params}")
        # Example logic to handle table creation, data insertion, and index creation
        if sql.startswith("CREATE TABLE"):
            table_name = sql.split()[2].strip('"')
            if table_name not in self.connection.root:
                self.connection.root[table_name] = OOBTree()
                transaction.commit()
                self.rowcount = 0
            else:
                print(f"Table {table_name} already exists")
        elif sql.startswith("INSERT INTO"):
            table_name = sql.split()[2].strip('"')
            table = self.connection.root.get(table_name)
            if table is not None:
                row_id = len(table) + 1
                table[row_id] = params
                transaction.commit()
                self.lastrowid = row_id
                self.rowcount = 1
            else:
                raise DatabaseError(f"Table {table_name} does not exist")
        elif sql.startswith("CREATE INDEX"):
            parts = sql.split()
            index_name = parts[2].strip('"')
            table_name = parts[4].strip('"')
            column_name = parts[5].strip('()"')
            catalog = self.connection.root['catalog']
            if index_name not in catalog:
                catalog[index_name] = FieldIndex(column_name)
                transaction.commit()
                self.rowcount = 0
            else:
                print(f"Index {index_name} already exists")
        else:
            print("Unsupported SQL command")

    def fetchall(self):
        # Implement fetching data from ZODB here
        return []

    def fetchmany(self, size=None):
        # Implement fetching data from ZODB here
        return []

    def close(self):
        print("Closing cursor")

    def __enter__(self):
        return self

    def __exit__(self, exc_type, exc_val, exc_tb):
        self.close()
