pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools"]

[project]
name = "django-zodb"
dependencies = ["django", "zodb", "zeo", "zope.catalog"]
dynamic = ["version"]

[tool.setuptools]
packages = ["django_zodb"]

[tool.setuptools.dynamic]
version = {attr = "django_zodb.__version__"}
